refactor: move modelling logic to separate module
