+++ iati_tables/database.py
@@ -1,7 +1,11 @@
+import logging
 import os
 from typing import Any
 
-from sqlalchemy import Engine, create_engine
+from sqlalchemy import Engine, create_engine, text
+
+logger = logging.getLogger(__name__)
+
 
 schema = os.environ.get("IATI_TABLES_SCHEMA")
 
@@ -45,3 +49,22 @@ def create_field_sql(object_details, sqlite=False):
         fields_with_type.append(field)
 
     return ", ".join(fields), ", ".join(fields_with_type)
+
+
+def _create_table(table, con, sql, **params):
+    logger.debug(f"Creating table: {table}")
+    con.execute(
+        text(
+            f"""
+            DROP TABLE IF EXISTS "{table}";
+            CREATE TABLE "{table}" AS {sql};
+            """
+        ),
+        {**params},
+    )
+
+
+def create_table(table, sql, **params):
+    engine = get_engine()
+    with engine.begin() as con:
+        _create_table(table.lower(), con, sql, **params)
